Ultimos cambios

+++ src/app/api/todos/route.ts
@@ -1,4 +1,5 @@
 
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
@@ -39,10 +40,17 @@ export async function GET(request: Request) {
 
   export async function POST(request: Request) { 
 
+    const user = await getUserSessionServer();
+  
+    if ( !user ) {
+      return NextResponse.json('No autorizado',{ status: 401 });
+    }
+  
+  
     try {
       const { complete, description } = await postSchema.validate( await request.json() );
   
-      const todo = await prisma.todo.create({ data: { complete, description } })
+      const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } })
     
       
       return NextResponse.json(todo);
@@ -50,16 +58,21 @@ export async function GET(request: Request) {
     } catch (error) {
       return NextResponse.json( error, { status: 400 } );
     }
-}
-
-export async function DELETE(request: Request) { 
-
-  try {
-
-    await prisma.todo.deleteMany({ where: { complete: true } });
-    return NextResponse.json('Borrados');
-    
-  } catch (error) {
-    return NextResponse.json( error, { status: 400 } );
   }
-}
+  export async function DELETE(request: Request) { 
+
+    const user = await getUserSessionServer();
+  
+    if ( !user ) {
+      return NextResponse.json('No autorizado',{ status: 401 });
+    }
+  
+    try {
+  
+      await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
+      return NextResponse.json('Borrados');
+      
+    } catch (error) {
+      return NextResponse.json( error, { status: 400 } );
+    }
+  }
